#24 - moved assertWithin to TestUtil
